+++ folha-pagamento/src/main/java/com/payroll/model/Position.java
@@ -9,13 +9,12 @@ import jakarta.persistence.Id;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-
 @Entity
 public class Position {
 	
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idCargo;
+    private Long idCargo;
 	
 	@NotBlank
     private String nome;
@@ -26,7 +25,7 @@ public class Position {
     // Construtores
     public Position() {}
 
-    public Position(int idCargo, String nome, BigDecimal salarioBase) {
+    public Position(Long idCargo, String nome, BigDecimal salarioBase) {
         this.idCargo = idCargo;
         this.nome = nome;
         this.salarioBase = salarioBase;
@@ -37,11 +36,11 @@ public class Position {
     }
 
     // Getters e Setters
-    public int getIdCargo() {
+    public Long getIdCargo() {
         return idCargo;
     }
 
-    public void setIdCargo(int idCargo) {
+    public void setIdCargo(Long idCargo) {
         this.idCargo = idCargo;
     }
